+++ udata_front/frontend/helpers.py
@@ -17,7 +17,7 @@ from . import front
 from udata.core.dataset.apiv2 import dataset_fields
 from udata.core.dataset.models import Dataset
 from udata.models import db
-from udata.i18n import format_date, _, pgettext
+from udata.i18n import format_date, format_timedelta, _, pgettext
 from udata.search.result import SearchResult
 from udata.utils import camel_to_lodash
 from udata_front.theme import theme_static_with_version
@@ -322,6 +322,30 @@ def daterange(value, details=False):
         end = value.end.strftime(date_format)
 
     return '{start!s}–{end!s}'.format(start=start, end=end) if end else start
+
+
+def format_from_now(value):
+    '''
+    Format date as relative from now.
+    It displays "today" or format_timedelta content, based on date.
+    '''
+    today = date.today()
+    value_without_time = value.date()
+    if(value_without_time == today):
+        return _("today")
+    return format_timedelta(value_without_time - today, add_direction=True)
+
+
+@front.app_template_filter()
+def format_based_on_date(value):
+    '''
+    Format date relative form now if date is less than a month ago.
+    Otherwise, show a formatted date.
+    '''
+    delta = date.today() - value.date()
+    if(delta.days > 30):
+        return _("on %(date)s", date=format_date(value, "long"))
+    return format_from_now(value)
 
 
 @front.app_template_filter()