+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - New organization page [#230](https://github.com/etalab/udata-front/pull/230)[#233](https://github.com/etalab/udata-front/pull/233)
 - Align reuse sidebar with image [#234](https://github.com/etalab/udata-front/pull/234)
+- Show relative date on dataset and resource cards [#231](https://github.com/etalab/udata-front/pull/231)
 
 ## 3.2.1 (2023-03-28)